requirements.txt
asn1crypto==0.24.0
attrdict==2.0.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
decorator==4.3.0
emoji==0.5.1
enum34==1.1.6
future==0.17.1
heroku==0.1.4
idna==2.7
ipaddress==1.0.22
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
mongoengine==0.16.2
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.0
pymongo==3.7.2
python-dateutil==1.5
python-telegram-bot==11.1.0
requests==2.20.1
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
